// back-office/src/app/openapi/model/userViewAdminDTO.ts
/**
 * FinalStake
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserViewAdminDTO { 
    createdAt: string;
    email: string;
    enabled: boolean;
    id: string;
    locked: boolean;
    roles: Array<string>;
    username: string;
}