// back-office/src/app/openapi/model/userUpdateAdminDTO.ts
/**
 * FinalStake
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserUpdateAdminDTO { 
    enabled: boolean;
    locked: boolean;
    roles: Array<string>;
}